+++ packages/ente-api/src/db/slot.repo.ts
@@ -81,6 +81,7 @@ export class SlotRepo {
     result.teacher =
       slot.teacher === null ? null : UserRepo.toDto(slot.teacher);
     result.student = UserRepo.toDto(slot.entry.student);
+    result.forSchool = slot.entry.forSchool;
 
     return result;
   }